versoview_messages
Enum ToVersoMessage

#[non_exhaustive]
pub enum ToVersoMessage {

    SetConfig(ConfigFromController),
    Exit,
    ListenToOnCloseRequested,
    NavigateTo(Url),
    ListenToOnNavigationStarting,
    OnNavigationStartingResponse(Vec<u8>, bool),
    ExecuteScript(String),
    ListenToWebResourceRequests,
    WebResourceRequestResponse(WebResourceRequestResponse),
    SetSize(Size),
    SetPosition(Position),
    SetMaximized(bool),
    SetMinimized(bool),
    SetFullscreen(bool),
    SetVisible(bool),
    StartDragging,
    GetSize(Uuid),
    GetPosition(Uuid),
    GetMaximized(Uuid),
    GetMinimized(Uuid),
    GetFullscreen(Uuid),
    GetVisible(Uuid),
    GetScaleFactor(Uuid),
    GetCurrentUrl(Uuid),

}

Message sent from the controller to versoview

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

SetConfig(ConfigFromController)

Initial configs for versoview this will be the first message sent to Verso once we received the sender from ToControllerMessage::SetToVersoSender

Exit

Exit

ListenToOnCloseRequested

Register a listener on versoview for getting notified on close requested from the OS, veroview will send a ToControllerMessage::OnCloseRequested when that happens

NavigateTo(Url)

Navigate to this URL

ListenToOnNavigationStarting

Register a listener on versoview for getting notified on navigation starting, veroview will send a ToControllerMessage::OnNavigationStarting when that happens

OnNavigationStartingResponse(Vec<u8>, bool)

Response to a ToControllerMessage::OnNavigationStarting message from versoview

ExecuteScript(String)

Execute JavaScript

ListenToWebResourceRequests

Register a listener on versoview for getting notified on web resource requests

WebResourceRequestResponse(WebResourceRequestResponse)

Response to a ToControllerMessage::OnWebResourceRequested message from versoview

SetSize(Size)

Sets the webview window’s size

SetPosition(Position)

Sets the webview window’s position

SetMaximized(bool)

Maximize or unmaximize the window

SetMinimized(bool)

Minimize or unminimize the window

SetFullscreen(bool)

Sets the window to fullscreen or back

SetVisible(bool)

Show or hide the window

StartDragging

Moves the window with the left mouse button until the button is released

GetSize(Uuid)

Get the window’s size, need a response with ToControllerMessage::GetSizeResponse

GetPosition(Uuid)

Get the window’s position, need a response with ToControllerMessage::GetPositionResponse

GetMaximized(Uuid)

Get if the window is currently maximized or not, need a response with ToControllerMessage::GetMaximizedResponse

GetMinimized(Uuid)

Get if the window is currently minimized or not, need a response with ToControllerMessage::GetMinimizedResponse

GetFullscreen(Uuid)

Get if the window is currently fullscreen or not, need a response with ToControllerMessage::GetFullscreenResponse

GetVisible(Uuid)

Get the visibility of the window, need a response with ToControllerMessage::GetVisibleResponse

GetScaleFactor(Uuid)

Get the scale factor of the window, need a response with ToControllerMessage::GetScaleFactorResponse

GetCurrentUrl(Uuid)

Get the current URL of the webview, need a response with ToControllerMessage::GetCurrentUrlResponse